requirements.txt
PyQt5
pyqtgraph
PyOpenGL
skyfield
matplotlib
pandas
lightstreamer-client-lib
requests
